refactor(is-encrypted-object): fix spelling error
