+++ lib/decrypt.js
@@ -4,13 +4,13 @@ var nativeCrypto = require('native-crypto')
 var Promise = require('lie')
 
 var ignore = require('./utils/ignore.js')
-var isEncrypedObject = require('./utils/is-encryped-object')
+var isEncryptedObject = require('./utils/is-encrypted-object')
 
 module.exports = decryptDoc
 
 // Mostly copied from https://github.com/calvinmetcalf/crypto-pouch/blob/master/index.js#L157
 function decryptDoc (key, doc) {
-  if (!isEncrypedObject(doc)) return Promise.resolve(doc)
+  if (!isEncryptedObject(doc)) return Promise.resolve(doc)
 
   var data = Buffer.from(doc.data, 'hex')
   var tag = Buffer.from(doc.tag, 'hex')
