+++ lib/event-handler.js
@@ -10,7 +10,7 @@
 
 var decryptOne = require('./helpers/decrypt-one')
 var updateListening = require('./utils/update-listening')
-var isEncrypedObject = require('./utils/is-encryped-object')
+var isEncryptedObject = require('./utils/is-encrypted-object')
 
 module.exports = eventHandler
 
@@ -52,7 +52,7 @@ function eventHandler (state, prefix, parentEmitter, emitter, api, isRoot) {
       return
     }
 
-    if (!isEncrypedObject(object)) {
+    if (!isEncryptedObject(object)) {
       return
     }
