+++ lib/utils/is-encrypted-object.js
